% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_txvis.R
\name{create_txvis}
\alias{create_txvis}
\title{Create a txvis object from raw data.}
\usage{
create_txvis(patient, treatment, start, end, date_format, ev_patient = NULL,
  events = NULL, event_date = NULL, event_end_date = NULL,
  ev_date_format = date_format)
}
\arguments{
\item{patient}{Patient encoding for treatments.}

\item{treatment}{Treatment encoding.}

\item{start}{Start date for treatment in a format defined using \code{date_format}.}

\item{end}{End date for treatment in a format defined using \code{date_format}.}

\item{date_format}{Date format, as a character string.  See \code{as.Date}.}

\item{ev_patient}{A patient encoding for each individual event.}

\item{events}{Any point event in the treatment of each individual.}

\item{event_date}{The date of event, formatted as \code{date_format}.}

\item{event_end_date}{The date that event finished, formatted as \code{date_format}.}

\item{ev_date_format}{Date format for event dates, as a character string. See \code{as.Date}.}
}
\value{
This command returns an object of class \code{txvis}.
}
\description{
The \code{txvis} object contains encoding for patient, treatment and dates, and can have
encoding for additional discrete events.  The class is used throughout the package.
}
\examples{

 # `treat` is bundled with the package.
 
 hlth_data <- create_txvis(patient        = treat$pat_id, 
                           treatment      = treat$treatment,
                           start          = treat$start,
                           end            = treat$end,
                           date_format    = "\%d\%B\%Y",
                           ev_patient     = events$pat_id,
                           events         = events$event,
                           event_date     = events$start,
                           event_end_date = events$end)
                           
 hlth_data


}
\author{
Ellen Korol \email{ellen.korol@iconplc.com}
}
\references{
txvis Package: https://github.com/johnwilsonICON/txvis
}
\keyword{utilities}

