% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\docType{data}
\name{events}
\alias{events}
\title{Dummy events dataset.}
\format{A \code{data.frame} with 1800 rows and 4 columns.
\describe{
  \item{pat_id}{}
  \item{event}{}
  \item{start}{}
  \item{end}{}
}}
\usage{
events
}
\description{
A dataset containing unique events for 865 unique (simulated) patients, consisting of
1800 unique events, coded as either point events or events that span some period of time.
}
\keyword{datasets}

