% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reform_dates.R
\name{reform_dates}
\alias{reform_dates}
\title{Reformat dates from long to wide.}
\usage{
reform_dates(x, start = NULL, end = NULL, interval = "month",
  conflict = "majority")
}
\arguments{
\item{x}{A \code{txVis} object}

\item{start}{The starting period for the date matrix. Default is the earliest date in the \code{txVis} object.}

\item{end}{The ending period for the date matrix. Default is the earliest date in the \code{txVis} object.}

\item{interval}{The length of the interval in text format.  See Details.}

\item{conflict}{Conflict resolution.}
}
\description{
This is, broadly speaking a helper function, that turns the \code{txVis} object from a long
table to a wide table with rows for each patient
and columns for regular intervals of dates from a defined \code{start} and \code{end} along
a regular \code{interval}.  In cases where treatments overlap within a time bin there are 
three coded \code{conflict} resolution methods:
}
\details{
\code{majority} in which coding priority is given to the treatment that occupies the most time within the interval
\code{first} in which coding priority is given to the treatment that occurs first chronologically within the interval
\code{last} in which coding priority is given to the treatment that occurs last chronologically within the interval
}
\examples{

 hlth_data <- create_txVis(patient   = treat$patient, 
                           treatment = treat$treatment,
                           start     = treat$start,
                           end       = treat$end,
                           date_format = "\%B \%d, \%Y")
                           
 wide <- reform_dates(hlth_data)

}

