% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reform_nodes.r
\name{reform_nodes}
\alias{reform_nodes}
\title{A helper function.}
\usage{
reform_nodes(x, y)
}
\arguments{
\item{x}{A \code{txVis} object}

\item{y}{A treatment sequence.}
}
\description{
A helper function.
}

