% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_treat.R
\docType{data}
\name{treatments}
\alias{treatments}
\title{Generate treatment data.}
\format{An object of class \code{numeric} of length 8.}
\usage{
treatments
}
\description{
Generate treatment data.
}
\keyword{datasets}

