% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_d3alluv.r
\name{tx_d3alluvial}
\alias{tx_d3alluvial}
\title{Generate an alluvial plot for treatment data using D3.js.}
\usage{
tx_d3alluvial(txvis, nsequ = NULL, seq.v.dat = "seq", start = NULL,
  end = NULL, interval = "month", conflict = "majority")
}
\arguments{
\item{txvis}{An object of class \code{txvis}.}
}
\description{
Using a \code{txvis} object, plot the sequencing of treatments using an alluvial plot.
}

