% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_txVis.R
\name{create_txVis}
\alias{create_txVis}
\title{Create a txVis object from raw data.}
\usage{
create_txVis(patient, treatment, start, end, date_format, ev_patient = NULL,
  events = NULL, event_date = NULL)
}
\arguments{
\item{patient}{Patient encoding for treatments.}

\item{treatment}{Treatment encoding.}

\item{start}{Start date for treatment in a format defined using \code{date_format}.}

\item{end}{End date for treatment in a format defined using \code{date_format}.}

\item{date_format}{Date format, as a character string.  See \code{as.Date}.}

\item{events}{Any point event in the treatment of each individual.}

\item{event_date}{The date, formatted as \code{start} or \code{end}.}

\item{ev.patient}{A patient encoding for each individual event.}
}
\value{
This command returns an object of class \code{txVis}.
}
\description{
The \code{txVis} object contains encoding for patient, treatment and dates, and can have
encoding for additional discrete events.  The class is used throughout the package.
}
\examples{
\dontrun{

 # `treat` is bundled with the package.
 
 hlth_data <- create_txVis(patient   = treat$patient, 
                           treatment = treat$treatment,
                           start     = treat$start,
                           end       = treat$end,
                           date_format = "\%B \%d, \%Y")
                           
 hlth_data

}
}
\author{
Ellen Korol \email{ellen.korol@iconplc.com}
}
\references{
txVis Package: https://github.com/johnwilsonICON/TxVis
}
\keyword{utilities}

