% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_alluv.r
\name{tx_alluvial}
\alias{tx_alluvial}
\title{Generate an alluvial plot for treatment data.}
\usage{
tx_alluvial(txvis, nsequ = NULL, seq.v.dat = "seq", start = NULL,
  end = NULL, interval = "month", conflict = "majority",
  tx_colour = NULL, tx_cw = 0.05)
}
\arguments{
\item{txvis}{An object of class \code{txvis}.}

\item{nsequ}{The maximum number of sequences to plot.}
}
\description{
Using a txvis object, plot the sequencing of treatments using an alluvial plot.
}

