% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_alluv.r
\name{tx_alluvial}
\alias{tx_alluvial}
\title{Generate an alluvial plot for treatment data.}
\usage{
tx_alluvial(txVis, nsequ = NULL)
}
\arguments{
\item{txVis}{An object of class \code{txVis}.}

\item{nsequ}{The maximum number of sequences to plot.}
}
\description{
Using a txVis object, plot the sequencing of treatments using an alluvial plot.
}

