% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_icicle.r
\name{tx_icicle}
\alias{tx_icicle}
\title{Generate an icicle plot for treatment data.}
\usage{
tx_icicle(txvis, nsequ = NULL, seq.v.dat = "seq", start = NULL,
  end = NULL, interval = "month", conflict = "majority",
  tx_colour = NULL)
}
\arguments{
\item{txvis}{An object of class \code{txvis}.}
}
\description{
Using a \code{txvis} object, plot the sequencing of treatments using an icicle plot.
}

