% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_indivplot.r
\name{tx_indiv}
\alias{tx_indiv}
\title{Generate an icicle plot for treatment data.}
\usage{
tx_indiv(txvis, nsample = NULL, aligned = FALSE, clustered = FALSE,
  events = FALSE)
}
\arguments{
\item{txvis}{An object of class \code{txvis}.}

\item{nsample}{The number of patients to show sequence data for.  Default is 10.}

\item{aligned}{Should treatment sequences be displayed by date, or from a common origin.  Default is FALSE.}

\item{clustered}{Organize treatments so that similar treatments are adjacent to one another.  Not implemented.}

\item{events}{Should individual events be superimposed onto the treatment sequences. Defaults to FALSE.}
}
\value{
Returns a \code{ggplot2} object.
}
\description{
Using a \code{txvis} object, plot the sequencing of treatments using an icicle plot.
}
\examples{

 hlth_data <- create_txvis(patient        = treat$pat_id, 
                           treatment      = treat$treatment,
                           start          = treat$start,
                           end            = treat$end,
                           date_format    = "\%d\%b\%Y",
                           ev_patient     = events$pat_id,
                           events         = events$event,
                           event_date     = events$start,
                           event_end_date = events$end)
                           
 # Plot without events:                          
 tx_indiv(hlth_data)
 
 # Plot with event data:
 tx_indiv(hlth_data, events = TRUE)
 
 #  Add additional ggplot2 styline:
 library(ggplot2)
 
 tx_indiv(hlth_data) + theme_bw()
 
 # Show many more samples & events, and assign a common start date.
 tx_indiv(hlth_data, nsample = 50, aligned = TRUE) +
   theme(axis.text.y = element_text(size = 8))
 

}

