% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_sunburst.r
\name{tx_sunburst}
\alias{tx_sunburst}
\title{Generate an sunburst plot for treatment data.}
\usage{
tx_sunburst(txVis, nsequ = NULL, seq.v.dat = "seq", start = NULL,
  end = NULL, interval = "month", conflict = "majority",
  tx_colour = NULL)
}
\arguments{
\item{txVis}{An object of class \code{txVis}.}
}
\description{
Using a \code{txVis} object, plot the sequencing of treatments using a sunburst plot.
}

