% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_transmat.r
\name{tx_transmat}
\alias{tx_transmat}
\title{Generate a transition matrix.}
\usage{
tx_transmat(txVis)
}
\arguments{
\item{txVis}{An object of class \code{txVis}.}
}
\description{
I'm not quite sure why we have this.
}

